WIP: Loop over CR calls
- that works out of the box

Changes to be committed:
	modified:   syncopy/statistics/compRoutines.py
	modified:   syncopy/statistics/jackknifing.py
	modified:   syncopy/statistics/summary_stats.py
