+++ syncopy/statistics/compRoutines.py
@@ -17,6 +17,122 @@ from syncopy.shared.kwarg_decorators import process_io
 
 
 @process_io
+def npstats_cF(trl_dat, operation='mean', axis=0, noCompute=False, chunkShape=None):
+
+    """
+    Numpy summary statistics on single-trial arrays along indicated `axis`.
+
+    Parameters
+    ----------
+    trl_dat : :class:`numpy.ndarray`
+        Single trial data of arbitrary dimension
+    operation : {'mean', 'std', 'var', 'median'}
+        The statistical operation to perform
+    axis : int
+        The axis over which to calulate the average
+
+    Returns
+    -------
+    res : :class:`numpy.ndarray`
+        Result of the average
+
+    Notes
+    -----
+    This method is intended to be used as :meth:`~syncopy.shared.computational_routine.ComputationalRoutine.computeFunction`
+    inside a :class:`~syncopy.shared.computational_routine.ComputationalRoutine`.
+    Thus, input parameters are presumed to be forwarded from a parent metafunction.
+    Consequently, this function does **not** perform any error checking and operates
+    under the assumption that all inputs have been externally validated and cross-checked.
+    """
+
+    if noCompute:
+        # initialize result array
+        out_shape = list(trl_dat.shape)
+        out_shape[axis] = 1   # this axis will be summed over
+
+        return out_shape, trl_dat.dtype
+
+    return NumpyStatDim.methods[operation](trl_dat, axis=axis, keepdims=True)
+
+
+class NumpyStatDim(ComputationalRoutine):
+
+    """
+    Simple compute class which applies basic numpy statistical funtions
+    along  a ``dim`` of a Syncopy data object. The resulting Syncopy object
+    is of the same type as the input object, having one of its
+    dimensions reduced to a singleton due to the summary statistical operation.
+
+    Notes
+    -----
+    If ``keeptrials`` is set to False, a trial average is additionally computed
+    as in any other CR. For a standalone trial average use the _trial_statistics function.
+
+    See also
+    --------
+    _trial_statistics: Sequential computation of statistics over trials
+    """
+
+    methods = {'mean': np.nanmean,
+               'std': np.nanstd,
+               'var': np.nanvar,
+               'median': np.nanmedian
+               }
+
+    computeFunction = staticmethod(npstats_cF)
+
+    def process_metadata(self, in_data, out_data):
+
+        # the dimension over which the statistic got computed
+        dim = in_data.dimord[self.cfg['axis']]
+
+        out_data.samplerate = in_data.samplerate
+
+        # Get/set timing-related selection modifiers
+        # We've set a fallback all-to-all selection in any case
+
+        # time axis really gone, only one trial and time got averaged out
+        if dim == 'time' and not self.keeptrials:
+            trldef = np.array([[0, 1, 0]])
+
+        # trial average, needs equal trial lengths.. just copy from 1st
+        elif dim != 'time' and not self.keeptrials:
+            trldef = in_data.selection.trialdefinition[0, :][None, :]
+
+        # each trial has empty time axis, so we attach trivial trialdefinition:
+        # 1 sample per trial for stacking
+        elif dim == 'time' and self.keeptrials:
+            nTrials = len(in_data.selection.trials)
+            stacking_time = np.arange(nTrials)[:, None]
+            trldef = np.hstack((stacking_time, stacking_time + 1,
+                               np.zeros((nTrials, 1))))
+
+        # nothing happened on the time axis
+        else:
+            trldef = in_data.selection.trialdefinition
+
+        out_data.trialdefinition = trldef
+
+        # Get/set dimensional attributes changed by selection
+        for prop in in_data.selection._dimProps:
+            selection = getattr(in_data.selection, prop)
+            # due to fallback all-to-all selection this captures
+            # all existing dimensions
+            if selection is not None:
+                # propagate without change
+                if dim not in prop:
+                    if np.issubdtype(type(selection), np.number):
+                        selection = [selection]
+                    setattr(out_data, prop, getattr(in_data, prop)[selection])
+                # set to singleton or None
+                else:
+                    # numerical freq axis is gone after averaging
+                    if dim == 'freq':
+                        out_data.freq = None
+                    else:
+                        setattr(out_data, prop, [self.cfg['operation']])
+
+@process_io
 def cov_cF(trl_dat,
            ddof=None,
            statAxis=0,
@@ -109,7 +225,6 @@ class Covariance(ComputationalRoutine):
         out.samplerate = data.samplerate
         out.channel_i = np.array(data.channel[chanSec])
         out.channel_j = np.array(data.channel[chanSec])
-        
 
 
 @process_io
