+++ syncopy/statistics/jackknifing.py
@@ -9,48 +9,42 @@ from copy import deepcopy
 # Syncopy imports
 import syncopy as spy
 from syncopy.shared.parsers import data_parser, scalar_parser, array_parser
-from syncopy.shared.tools import get_defaults, get_frontend_cfg
-from syncopy.datatype import TimeLockData
 
 from syncopy.shared.errors import SPYValueError, SPYTypeError, SPYInfo
 from syncopy.shared.kwarg_decorators import (
     unwrap_cfg,
-    unwrap_select,
-    detect_parallel_client
+    unwrap_select
 )
-from syncopy.shared.input_processors import check_passed_kwargs
-from syncopy.shared.latency import get_analysis_window, create_trial_selection
 
-# Local imports
-from syncopy.statistics.compRoutines import PSTH
+from syncopy.statistics.compRoutines import NumpyStatDim
 from syncopy.statistics.psth import Rice_rule, sqrt_rule, get_chan_unit_combs
 
-available_binsizes = {'rice': Rice_rule, 'sqrt': sqrt_rule}
-available_outputs = ['rate', 'spikecount', 'proportion']
 
 # create test data on the fly
+ad = spy.AnalogData(data=[np.ones((10,4)) for _ in range(10)], samplerate=1)
+axis = ad.dimord.index('time')
+CR = NumpyStatDim(operation='mean', axis=axis)
 
 @unwrap_select
-def jacknife_cr(data, CR, CR_kwargs, **kwargs):
+def jacknife_cr(spy_data, CR, **kwargs):
     """
-    General meta-function to compute the jackknife estimate
+    General meta-function to compute the trial jackknife estimates
     of an arbitrary ComputationalRoutine by creating
-    leave-one-out selections
+    the full set of leave-one-out (loo) selections.
     """
-    print(data.selection, 's')
 
-    if data.selection is not None:
+    if spy_data.selection is not None:
         # create a back up
-        selection_backup = deepcopy(data.selection.select)
+        select_backup = deepcopy(spy_data.selection.select)
         selection_cleanup = False
     else:
         # create all-to-all selection
         # for easier property propagation
-        data.selectdata(inplace=True)
+        spy_data.selectdata(inplace=True)
         selection_cleanup = True
 
     # now we have definitely a selection
-    all_trials = data.selection.trial_ids
+    all_trials = spy_data.selection.trial_ids
 
     # collect the leave-one-out (loo) trial selections
     loo_trial_selections = []
@@ -60,17 +54,36 @@ def jacknife_cr(data, CR, CR_kwargs, **kwargs):
         loo.remove(trl_id)
         loo_trial_selections.append(loo)
 
-    # strip of trial selection at first
-    select = data.selection.select
-    select.pop('trials')
+    # --- CR computations --
+
+    log_dict = {}
+
+    # initialize jackknife output object of same datatype
+    jack_out = spy_data.__class__(dimord=spy_data.dimord)
+
+    # manipulate existing selection
+    select = spy_data.selection.select
     # create loo selections and run CR
     # to compute jackknife replicates
     for loo in loo_trial_selections:
         select['trials'] = loo
-        data.selectdata(select, inplace=True)
-        print(data.selection)
+        spy_data.selectdata(select, inplace=True)
+
+        # initialize transient output object of same datatype
+        out = spy_data.__class__(dimord=spy_data.dimord)
+
+        # (re-)initialize supplied CR and compute a trial average
+        CR.initialize(spy_data, spy_data._stackingDim,
+                      keeptrials=False, chan_per_worker=kwargs.get('chan_per_worker'))
+        CR.compute(spy_data, out, parallel=kwargs.get("parallel"), log_dict=log_dict)
+
+        # each replicate gets trial averaged, grab that trial and write
+        # into jackknife output
+        out.trials[0]
 
     if selection_cleanup:
-        data.selection = None
+        spy_data.selection = None
+    else:
+        spy_data.selectdata(select_backup)
 
     return loo_trial_selections
